% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{elastic_aipw_ses_l2_list}
\alias{elastic_aipw_ses_l2_list}
\title{Simulation results for beta0 = c(0,1,1,1) and psi0 = c(0,1,1) in Yang et al., (2022), Section S4.1
with median separation}
\format{
\code{elastic_aipw_ses_l2_list}
A data list with no unmeasured confounding
}
\usage{
elastic_aipw_ses_l2_list
}
\description{
Simulation results for beta0 = c(0,1,1,1) and psi0 = c(0,1,1) in Yang et al., (2022), Section S4.1
with median separation
}
\keyword{datasets}
