% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{elastic_psi000_fixed_lists}
\alias{elastic_psi000_fixed_lists}
\title{Simulation results for beta0 = c(0,1,1,1) and psi0 = c(0,0,0) in Yang et al., (2022), Section S4.5}
\format{
\code{elastic_psi000_fixed_lists}
A data list with length 10,
each one associated with different strength of unmeasured confounding (e.g., 0, ..., 2)
}
\usage{
elastic_psi000_fixed_lists
}
\description{
Simulation results for beta0 = c(0,1,1,1) and psi0 = c(0,0,0) in Yang et al., (2022), Section S4.5
}
\keyword{datasets}
