% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elastic_psi011_lists}
\alias{elastic_psi011_lists}
\title{Simulation results for beta0 = c(0,1,1,1) and psi0 = c(0,1,1) in Yang et al., (2022), Section 4}
\format{
##\code{elastic_psi011_lists}
A data list with length 10,
each one associated with different strength of unmeasured confounding (e.g., 0, ..., 2)
}
\usage{
elastic_psi011_lists
}
\description{
Simulation results for beta0 = c(0,1,1,1) and psi0 = c(0,1,1) in Yang et al., (2022), Section 4
}
\keyword{datasets}
