% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticHTE.R
\name{elasticHTE}
\alias{elasticHTE}
\title{Elastic integrative analysis for heterogenous treatment effect}
\usage{
elasticHTE(dat.t, dat.os, fixed = FALSE)
}
\arguments{
\item{dat.t}{A list contains X, Y and A for RCT.}

\item{dat.os}{A list contains X, Y and A fro RWE.}

\item{fixed}{A boolean regarding how to select the tuning parameter \code{c_gamma}.
\itemize{
\item \code{FALSE} (the default): use adaptive selection strategy
\item \code{TRUE}: use fixed threshold strategy.
The default fixed threshold is \link{qchisq(1 -0.05, df = p)},
in which \code{p} is the dimension of \code{X} plus one (i.e., add the intercept).
}}
}
\value{
A list with components:
\itemize{
\item est: estimated \code{psi} associated with the treatment modifiers.
\item ve: estimated standard error for \code{psi}.
\item CIs.inf, CIs.sup: estimated confidence intervals for \code{psi}.
\item nuispar: a vector of estimated nuisance parameters
}
}
\description{
\code{elasticHTE()} is a test-based dynamic borrowing framework combining
a randomized clinical trial (RCT) and a real-world evidence (RWE) study,
in which a preliminary test statistic is used to gauge the comparability
and reliability of the RWE and decide whether or not to use the RWE in
an integrative analysis. The parameter of interest is \code{psi}, which quantifies
how the treatment effect varies over the treatment modifiers.
}
\examples{
m <- 2000 # size of RW dataset
# generate the data
datasets.all <- GenerateData(psi0 = c(0,1,1),
                             tlocalpar = 2,
                             m = m)
# choice of kappa_n, default is tau * sqrt(log(m)) (similar to BIC criteria)
thres.psi  <-  sqrt(log(m)) # threshold for ECI psi
# conduct the elastic integrative analysis
result.elastic <- ElasticEst(dat.t = datasets.all$RT,
                             dat.os = datasets.all$RW)
}
