% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.res}
\alias{plot.res}
\title{ggplot for plotting results}
\usage{
\method{plot}{res}(x, AIPW = T)
}
\arguments{
\item{x}{an list of class "res".}

\item{AIPW}{logical. If \code{TRUE}, the AIPW estimator will not be plotted.}
}
\description{
\code{plt.res.combine()} reproduce the plots in Yang et al., (2022), Figure 4 and Figure S2.
}
