% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ggplot.R
\name{plt.res.combine}
\alias{plt.res.combine}
\title{ggplot for plotting}
\usage{
plt.res.combine(bias.psi0, bias.psi1, variance.psi0, variance.psi1, AIPW = T)
}
\description{
\code{plt.res.combine()} reproduce the plots in Yang et al., (2022), Figure 4
}
