% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.res}
\alias{summary.res}
\title{summarize the results}
\usage{
\method{summary}{res}(res, psi)
}
\arguments{
\item{res}{a list of class "res", produced by data-raw/DATASET_SIM.R or data-raw/DATASET_AIPW.R.}

\item{psi}{a vector of \code{psi}, useful for computing the bias and mean squared error.}
}
\description{
\code{summary.res()} summarizes the simulation outputs in Yang et al., (2022), Table 1
}
