% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summary.tables}
\alias{summary.tables}
\title{Summary results for beta0 = c(0,1,1,1) and psi0 = c(0,1,1) in Yang et al., (2022), Table 1}
\format{
##\code{summary.tables}
A data frame with the coverage rates and confidence intervals
with different strength of unmeasured confounding
}
\usage{
summary.tables
}
\description{
Summary results for beta0 = c(0,1,1,1) and psi0 = c(0,1,1) in Yang et al., (2022), Table 1
}
\keyword{datasets}
