% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ggplot.R
\name{summary_res}
\alias{summary_res}
\title{summarize the results}
\usage{
summary_res(res, psi)
}
\description{
\code{summary_res()} summarizes the simulation outputs in Yang et al., (2022), Table 1
}
