% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticHTE.R
\name{elasticHTE}
\alias{elasticHTE}
\title{Elastic integrative analysis for heterogenous treatment effect}
\usage{
elasticHTE(
  mainName,
  contName,
  propenName,
  dat.t,
  dat.os,
  sl.lib.A = c("SL.glm"),
  sl.lib.Y = c("SL.glmnet"),
  thres.psi = sqrt(log(length(dat.os$Y))),
  fixed = FALSE,
  alpha = rep(0, length(contName) + 1),
  ...
)
}
\arguments{
\item{mainName}{A character object dictates the covariates used to model the main effects of the outcomes}

\item{contName}{A character object dictates the covariates used to model the contrasts of the outcomes}

\item{propenName}{A character object dictates the covariates used to model the propensities for treatment}

\item{dat.t}{A \code{\link[=data.frame]{data.frame()}} contains X, Y and A for RCT.}

\item{dat.os}{A \code{\link[=data.frame]{data.frame()}} contains X, Y and A for RWE.}

\item{sl.lib.A}{A character vector of prediction algorithms to model propensity scores for RWE (and RCT if not specified).
A list of available functions can be found with \code{\link[SuperLearner:listWrappers]{SuperLearner::listWrappers()}}.}

\item{sl.lib.Y}{A character vector of prediction algorithms to model conditional outcomes for RCT and RWE.}

\item{thres.psi}{A numerical value indicating the threshold for constructing adaptive confidence interval.}

\item{fixed}{A boolean regarding how to select the tuning parameter \code{c_gamma}.
\itemize{
\item \code{FALSE} (the default): use adaptive selection strategy
\item \code{TRUE}: use fixed threshold strategy.
The default fixed threshold is \link{qchisq(1 -0.05, df = p)},
in which \code{p} is the dimension of \code{X} plus one (i.e., add the intercept).
}}

\item{alpha}{a vector for estimating the linear combination of the coefficients for treatment modifiers (depreciated).}

\item{...}{additional arguments to be passed to \code{\link[SuperLearner:SuperLearner]{SuperLearner::SuperLearner()}}.}
}
\value{
A list with components:
\itemize{
\item est: estimated \code{psi} associated with the treatment modifiers.
\item ve: estimated standard error for \code{psi}.
\item CIs.inf, CIs.sup: estimated confidence intervals for \code{psi}.
\item nuispar: a vector of estimated nuisance parameters
}
}
\description{
\code{elasticHTE()} is a test-based dynamic borrowing framework combining
a randomized clinical trial (RCT) and a real-world evidence (RWE) study,
in which a preliminary test statistic is used to gauge the comparability
and reliability of the RWE and decide whether or not to use the RWE in
an integrative analysis. The parameter of interest is \code{psi}, which quantifies
how the treatment effect varies over the treatment modifiers.
}
\examples{
m <- 2000 # size of RW dataset
# generate the data
datasets.all <- GenData(psi0 = c(0,1,1),
                             tlocalpar = 2,
                             m = m)
# datasets.all$RT: the dataset for RT
# datasets.all$RW: the dataset for RW
# $Y: is a vector of observed outcome
# $X: is a matrix of covariates without intercept (n x (p-1))
# $A: is a vector of received treatment
# $q: is a vector of 1, for permutation-based variance estimation
# RT$ps, RT$ml.ps: is a vector of known propensity score (only available in RT)

# choice of kappa_n, default is tau * sqrt(log(m)) (similar to BIC criteria)
thres.psi  <-  sqrt(log(m)) # threshold for ECI psi
# conduct the elastic integrative analysis
result.elastic <- elasticHTE(mainName = c('X1', 'X2'),
                             contName = c('X1', 'X2'),
                             propenName = c('X1', 'X2'),
                             dat.t = datasets.all$RT,
                             dat.os = datasets.all$RW,
                             fixed = FALSE)
result.elastic
# return: a list called result.elastic
# $est: estimates of the psi in the finite population
# $ve: variance estimation of the psi by permutation-based estimation
# $CIs.inf, CIs.sup: 95\% confidence intervals
  ## Wald CIs for regular estimators;
  ## ACIs for non-regular estimators, namely elas.v1 and elas.v2
  ## elas.v1 performs better in our simulation studies.
# $nuispar: estimates for various nuisance parameter
}
